requirements.txt
django
djangorestframework
serverless-wsgi
whitenoise
